+++ src/Uploadcare/Uploader.php
@@ -56,12 +56,12 @@ class Uploader
    */
   public function fromUrl($url, $check_status = true, $timeout = 1, $max_attempts = 5)
   {
-    $data = array(
+    $requestData = array(
         '_' => time(),
         'source_url' => $url,
         'pub_key' => $this->api->getPublicKey(),
     );
-    $ch = $this->__initRequest('from_url', $data);
+    $ch = $this->__initRequest('from_url', $requestData);
     $this->__setHeaders($ch);
 
     $data = $this->__runRequest($ch);
@@ -232,7 +232,7 @@ class Uploader
   {
     curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
     curl_setopt($ch, CURLOPT_HTTPHEADER, array(
-      'User-Agent: ' . $this->api->ua,
+      'User-Agent: ' . $this->api->getUserAgent(),
     ));
   }