+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.5.1
+- add throttled requests handling
+- add customizable User Agent string
+
 ## 1.5.0
 - upgrade server api usage to v0.4, add new style pagination support 
 - add FileIterator